+++ src/main/java/com/fullstack/customer/CustomerDao.java
@@ -6,10 +6,10 @@ import java.util.stream.Stream;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomer();
-    Optional<Customer> selectById(Integer id);
+    Optional<Customer> selectById(Long id);
     void insertCustomer(Customer customer);
     boolean existCustomerByEmail(String email);
-    void deleteCustomer(Integer id);
-    boolean existsCustomerById(Integer id);
+    void deleteCustomer(Long  id);
+    boolean existsCustomerById(Long id);
     void updateCustomer(Customer update);
 }
